move roots to a module
